update_slackmoji/requirements.txt
requests
PyYAML
pillow
python-box
